Revert buffer overflow fix in PrefaceByArticle().  This caused problems with Parchment's caching.  Parchment has since been fixed but this code is excised anyway.  Andrew Plotkin described the alteration as follows:
It looks like this is altering (and then restoring) the contents of the
object property table. This sounds like a terrible idea on the face of
things, and the code looks buggy on top of that -- what if i-->1 is zero
to begin with? (It will be altered but not restored.) Why is this even
performing surgery on the object's hardware name, when PSN__(o) is more
likely to print the short_name property?