Homogenize the Prompt of AfterGameOver
The final prompt (following "would you like to restart", etc) was
printed with "> "; this has an extra space compared to the normal prompt
(and is not the normal prompt). Replacing this by L__M(##Prompt)
homogenizes things and gives an entry point for the author to have a
special prompt after the final message. It requires modifying Miscellany
5 so it doesn't skip a line there; this should make sense for an author
changing Misc5: the prompt should be on its own line, so that's one line
break, and if you return a string you add another one, so you should
print (and not return) if you only want one line break.

Another solution would be to modify L__M(##Prompt) so that it prints a
line break only if deadflag is 0. This will get called at every turn, so
it's less efficient.