Partial fix for #23 (Mantis 1813): Hangup between grammar and orders routines
Problem: Orders fully parsed by an animate/talkable object's grammar
routine are not being delivered to its orders routine.

Preliminary analysis: Prior to Roger Firth's movement of code from before
"! Generate the action" to InformLibrary.actor_act(), that code would
ALWAYS result in a jump to begin__action or to turn__end if (actor ~=
player).  After the change, this was not always the case.  It seems that
this change was made to accomodate the new "<action [noun] [second], actor>"
grammar for invoking R_Process (introduced in Inform compiler 6.33).
For the moment, I'm rolling back this chunk to prior to Roger's change
to have an anchor point from which to investigate a more perfect
solution.