Final fix for #23 (Mantis 1813): Two chunks of actor_act() consolidated.
The first time we call self.actor_act(), some things must be done
slightly differently in order to avoid problems when doing something to
several items at once.  For instance: TAKE ROCKS. See this function in
1d95759b1bd6674509d6cf9161e6db3da3831f10 to see how there used to be
essentially two functions.  The need for this distiction came about with
Issue #30.  So, if this is the first time calling actor_act(), pass true
on the fifth parameter (ft).