Issue 32: restore 6/11 version of L__M(##Drop, 2).
Restore lib 6/11 version of L__M(##Drop, 2) msg, updated for
voices and tenses. The msg was changed in response to
Suggestion 122, but the need no longer exists, because lib 6/12
performs an implicit take when an actor attempts to drop an
indirectly contained object. The msg is now only shown for
external objects where it is not always appropriate. The older,
more general msg is better in these cases.